@@ -19,28 +19,28 @@
     "lint": "npx nx run-many --target lint --all"
   },
   "dependencies": {
-    "@angular/animations": "^16",
-    "@angular/cdk": "^16.0.1",
-    "@angular/common": "^16",
-    "@angular/compiler": "^16",
-    "@angular/core": "^16",
+    "@angular/animations": "^16.1.6",
+    "@angular/cdk": "^16.1.5",
+    "@angular/common": "^16.1.6",
+    "@angular/compiler": "^16.1.6",
+    "@angular/core": "^16.1.6",
     "@angular/fire": "^7.6.1",
     "@angular/forms": "^16.1.6",
     "@angular/google-maps": "^16.1.5",
     "@angular/material": "^16.1.5",
-    "@angular/platform-browser": "^16",
+    "@angular/platform-browser": "^16.1.6",
     "@angular/platform-browser-dynamic": "^16.1.6",
     "@angular/platform-server": "^16.1.6",
     "@angular/router": "^16.1.6",
     "@nguniversal/express-engine": "^16.1.1",
-    "rxjs": "^7",
-    "tslib": "^2.6.0",
-    "zone.js": "~0.13.0"
+    "rxjs": "^7.8.1",
+    "tslib": "^2.6.1",
+    "zone.js": "^0.13.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^16.1.5",
-    "@angular-devkit/core": "^16",
-    "@angular-devkit/schematics": "^16",
+    "@angular-devkit/core": "^16.1.5",
+    "@angular-devkit/schematics": "^16.1.5",
     "@angular-eslint/eslint-plugin": "^16.1.0",
     "@angular-eslint/eslint-plugin-template": "^16.1.0",
     "@angular-eslint/template-parser": "^16.1.0",
@@ -48,32 +48,31 @@
     "@angular/compiler-cli": "^16.1.6",
     "@angular/language-service": "^16.1.6",
     "@nguniversal/builders": "^16.1.1",
-    "@nx/angular": "^16",
-    "@nx/esbuild": "16.5.5",
-    "@nx/eslint-plugin": "16.5.5",
-    "@nx/js": "16.5.5",
-    "@nx/linter": "16.5.5",
-    "@nx/workspace": "^16",
-    "@schematics/angular": "^16",
-    "@types/node": "18.7.1",
+    "@nx/angular": "^16.5.5",
+    "@nx/esbuild": "^16.5.5",
+    "@nx/eslint-plugin": "^16.5.5",
+    "@nx/js": "^16.5.5",
+    "@nx/linter": "^16.5.5",
+    "@nx/workspace": "^16.5.5",
+    "@schematics/angular": "^16.1.5",
+    "@types/node": "^18.17.1",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
-    "esbuild": "^0.17.17",
     "eslint": "^8.45.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-tsdoc": "^0.2.17",
     "firebase-admin": "^11.10.1",
     "firebase-functions": "^4.4.1",
-    "firebase-tools": "^12.4.5",
+    "firebase-tools": "^12.4.6",
     "git-describe": "^4.1.1",
-    "ng-packagr": "~16.1.0",
-    "nx": "^16",
-    "nx-cloud": "^16",
-    "postcss": "^8.4.5",
-    "postcss-import": "~14.1.0",
-    "postcss-preset-env": "~7.5.0",
-    "postcss-url": "~10.1.3",
+    "ng-packagr": "^16.1.0",
+    "nx": "^16.5.5",
+    "nx-cloud": "^16.1.1",
+    "postcss": "^8.4.27",
+    "postcss-import": "^14.1.0",
+    "postcss-preset-env": "^7.8.3",
+    "postcss-url": "^10.1.3",
     "prettier": "^2.8.8",
-    "typescript": "~5.1.3"
+    "typescript": "^5.1.6"
   }
 }
