@@ -1,5 +1,5 @@
-import { HttpsFunction, Request, Response, runWith } from "firebase-functions";
-import { FunctionRequest, FunctionResponse }         from "@heypoint/types";
+import { https, HttpsFunction, Response, runWith } from "firebase-functions";
+import { FunctionResponse }                        from "@heypoint/types";
 
 
 // noinspection JSUnusedGlobalSymbols
@@ -7,6 +7,6 @@ export const endpoint: HttpsFunction = runWith({
   enforceAppCheck: true,
 })
   .https
-  .onRequest((request: Request, response: Response<FunctionResponse>): void => response.send({
-    message: request.body
+  .onRequest((request: https.Request, response: Response<FunctionResponse>): void => response.send({
+    message: "hello world"
   }).end() && void(0));
